+++ pkg/beacon/relay/config/config.go
@@ -12,15 +12,15 @@ type Chain struct {
 	// TicketInitialSubmissionTimeout is the duration (in blocks) the staker has to submit
 	// tickets that fall under the natural threshold to satisfy the initial
 	// ticket timeout (see group selection, phase 2a).
-	TicketInitialSubmissionTimeout uint8
+	TicketInitialSubmissionTimeout uint64
 	// TicketReactiveSubmissionTimeout is the duration (in blocks) the staker has to
 	// submit any tickets that did not fall under the natural threshold. This
 	// final chance to submit tickets is called reactive ticket submission
 	// (defined in the group selection algorithm, 2b).
-	TicketReactiveSubmissionTimeout uint8
+	TicketReactiveSubmissionTimeout uint64
 	// TicketChallengeTimeout is the duration (in blocks) the staker has to
 	// submit any challenges for tickets that fail any checks.
-	TicketChallengeTimeout uint8
+	TicketChallengeTimeout uint64
 	// ResultPublicationBlockStep is the duration (in blocks) that has to pass
 	// before group member with the given index is eligible to submit the
 	// result.
@@ -28,7 +28,7 @@ type Chain struct {
 	// T_dkg + (N-1) * T_step
 	// where T_dkg is time for phases 1-12 to complete and T_step is the result
 	// publication block step.
-	ResultPublicationBlockStep uint8
+	ResultPublicationBlockStep uint64
 	// MinimumStake is an on-chain value representing the minimum necessary
 	// amount a client must lock up to submit a single ticket
 	MinimumStake *big.Int
